requirements.txt
# Core ML libraries
lightgbm>=4.0.0
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Experiment tracking
mlflow>=2.8.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# CLI
click>=8.1.0
tqdm>=4.65.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Data validation
pydantic>=2.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Utilities
joblib>=1.3.0

# Web App
streamlit>=1.30.0
plotly>=5.0.0
